[refactor] BoothDetailScreen 내에 Component 분리
