+++ feature/booth/src/main/kotlin/com/unifest/android/feature/booth/component/BoothDescription.kt
@@ -1,0 +1,118 @@
+package com.unifest.android.feature.booth.component
+
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.width
+import androidx.compose.foundation.layout.widthIn
+import androidx.compose.material3.Icon
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.remember
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.graphics.vector.ImageVector
+import androidx.compose.ui.layout.LastBaseline
+import androidx.compose.ui.platform.LocalConfiguration
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.res.vectorResource
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import com.unifest.android.core.designsystem.ComponentPreview
+import com.unifest.android.core.designsystem.R
+import com.unifest.android.core.designsystem.component.UnifestOutlinedButton
+import com.unifest.android.core.designsystem.theme.BoothCaution
+import com.unifest.android.core.designsystem.theme.BoothLocation
+import com.unifest.android.core.designsystem.theme.BoothTitle1
+import com.unifest.android.core.designsystem.theme.Content2
+import com.unifest.android.core.designsystem.theme.Title5
+import com.unifest.android.core.designsystem.theme.UnifestTheme
+import com.unifest.android.feature.booth.viewmodel.BoothUiAction
+
+@Composable
+fun BoothDescription(
+    name: String,
+    warning: String,
+    description: String,
+    location: String,
+    onAction: (BoothUiAction) -> Unit,
+) {
+    val configuration = LocalConfiguration.current
+    val maxWidth = remember(configuration) {
+        val screenWidth = configuration.screenWidthDp.dp - 40.dp
+        screenWidth * (2 / 3f)
+    }
+
+    Column(modifier = Modifier.padding(horizontal = 20.dp)) {
+        Row(verticalAlignment = Alignment.CenterVertically) {
+            Text(
+                text = name,
+                modifier = Modifier
+                    .widthIn(max = maxWidth)
+                    .alignBy(LastBaseline),
+                color = MaterialTheme.colorScheme.onBackground,
+                style = BoothTitle1,
+            )
+            Spacer(modifier = Modifier.width(5.dp))
+            Text(
+                text = warning,
+                modifier = Modifier.alignBy(LastBaseline),
+                style = BoothCaution,
+                color = MaterialTheme.colorScheme.primary,
+            )
+        }
+        Spacer(modifier = Modifier.height(15.dp))
+        Text(
+            text = description,
+            modifier = Modifier.padding(top = 8.dp),
+            color = MaterialTheme.colorScheme.onSecondary,
+            style = Content2.copy(lineHeight = 18.sp),
+        )
+        Spacer(modifier = Modifier.height(20.dp))
+        Row(
+            verticalAlignment = Alignment.CenterVertically,
+            modifier = Modifier.padding(top = 8.dp),
+        ) {
+            Icon(
+                imageVector = ImageVector.vectorResource(id = R.drawable.ic_location_green),
+                contentDescription = "location icon",
+                tint = Color.Unspecified,
+            )
+            Spacer(modifier = Modifier.width(8.dp))
+            Text(
+                text = location,
+                color = MaterialTheme.colorScheme.onBackground,
+                style = BoothLocation,
+            )
+        }
+        Spacer(modifier = Modifier.height(16.dp))
+        UnifestOutlinedButton(
+            onClick = { onAction(BoothUiAction.OnCheckLocationClick) },
+            modifier = Modifier.fillMaxWidth(),
+        ) {
+            Text(
+                text = stringResource(id = R.string.booth_check_locaiton),
+                style = Title5,
+            )
+        }
+    }
+}
+
+@ComponentPreview
+@Composable
+fun BoothDescriptionPreview() {
+    UnifestTheme {
+        BoothDescription(
+            name = "공대주점",
+            warning = "누구나 환영",
+            description = "컴퓨터 공학과와 물리학과가 함께하는 협동부스입니다. 방문자 이벤트로 무료 안주 하나씩 제공중이에요!!",
+            location = "공학관",
+            onAction = {},
+        )
+    }
+}
